vscode: licensing
