+++ LICENSE
@@ -4,10 +4,11 @@ Each file includes a header which describes the relevant license.
 In broad strokes:
   wake          LICENSE.Apache2   everything not in vendor/
 
-  BLAKE2        LICENSE.CC0       vendor/blake2/
-  gopt          LICENSE.TFL       vendor/gopt/
-  lemon         public domain     vendor/lemon/
-  SipHash       LICENSE.CC0       vendor/siphash/
-  Unicode data  LICENSE.UNICODE   vendor/utf8proc/utf8proc_data.c
-  utf8proc      LICENSE.MIT       vendor/utf8proc/utf8proc.[ch]
-  whereami      LICENSE.WTFPL     vendor/whereami/
+  BLAKE2                    LICENSE.CC0       vendor/blake2/
+  gopt                      LICENSE.TFL       vendor/gopt/
+  lemon                     public domain     vendor/lemon/
+  SipHash                   LICENSE.CC0       vendor/siphash/
+  Unicode data              LICENSE.UNICODE   vendor/utf8proc/utf8proc_data.c
+  utf8proc                  LICENSE.MIT       vendor/utf8proc/utf8proc.[ch]
+  whereami                  LICENSE.WTFPL     vendor/whereami/
+  vscode extension clients  LICENSE.MIT       extensions/vscode/lsp-client/src/vendor
