+++ extensions/vscode/.vscode/launch.json
@@ -1,4 +1,4 @@
-// A launch configuration that compiles the extension and then opens it inside a new window
+// A launch configuration that runs the extension inside a new window
 {
 	"version": "0.2.0",
 	"configurations": [
